@@ -4,15 +4,12 @@
   "description": "Remap CSS rules based on declaration value",
   "author": "silverwind",
   "repository": "silverwind/remap-css",
+  "type": "module",
+  "exports": "./index.js",
   "license": "BSD-2-Clause",
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
-  "keywords": [
-    "remap",
-    "css",
-    "theme"
-  ],
   "files": [
     "index.js"
   ],
@@ -41,11 +38,9 @@
   "devDependencies": {
     "eslint": "7.26.0",
     "eslint-config-silverwind": "33.0.0",
-    "jest": "26.6.3",
     "updates": "12.1.0",
-    "versions": "8.4.7"
-  },
-  "jest": {
-    "verbose": false
+    "versions": "8.4.7",
+    "vite": "3.2.4",
+    "vitest": "0.25.2"
   }
 }
